Update docs, moves stuff into common
